move Sample struct into its own module and add a test
